+++ handler/Gopkg.toml
@@ -31,7 +31,7 @@
 
 [[constraint]]
   name = "github.com/aws/aws-sdk-go"
-  version = "1.25.25"
+  version = "1.25.29"
 
 [[constraint]]
   name = "github.com/caarlos0/env"
